requirements.txt
pip==22.0.4
pandas==1.4.2
numpy==1.22.3
bs4==0.0.1
requests==2.27.1
selenium==4.1.5
regex==2022.4.24
glob2==0.7
mysql-connector-python==8.0.29
sqlalchemy==1.4.36
pymysql==1.0.2